+++ continual/convert.py
@@ -7,7 +7,7 @@ from torch import Tensor, nn
 
 from .container import Sequential
 from .conv import Conv1d, Conv2d, Conv3d
-from .interface import CoModule
+from .module import CoModule
 from .logging import getLogger
 from .pooling import (
     AdaptiveAvgPool2d,