+++ tests/continual/test_container.py
@@ -4,7 +4,7 @@ import torch
 from torch import nn
 
 import continual as co
-from continual.interface import TensorPlaceholder
+from continual.module import TensorPlaceholder
 
 torch.manual_seed(42)